feat(user): 重构用户相关数据库结构
- 重命名数据库文件，优化表结构设计
- 添加用户组、用户日志、第三方账号等新表
- 更新用户表字段，增加权限相关列- 调整地址表，使用行政代码关联地区信息

+++ mall-user/src/main/java/com/jiang/mall/domain/enums/LogStatus.java
@@ -31,11 +31,11 @@ public enum LogStatus {
 	SUCCESS_ADMIN_LOCK( 10, "管理员锁定用户成功"),
 	SUCCESS_UNLOCK( 11, "解锁用户成功");
 
-	private final byte value;
+	private final Integer value;
 	private final String name;
 
 	LogStatus(int value, String name) {
-		this.value = (byte)value;
+		this.value = value;
 		this.name = name;
 	}
